requirements.txt
# Core Agent Framework
pyautogen==0.2.20
openai>=1.3.0

# LLM & Vector Tooling
sentence-transformers==2.7.0
faiss-cpu==1.8.0
torch==2.3.0
transformers==4.41.2
scikit-learn==1.4.2

# API Server & Data Validation
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
requests==2.31.0

# Geospatial Calculations
geopy==2.4.1